Define `fpp_t` and `f_cnt_t` to be of `size_t` (#7363)
Defines `fpp_t` and `f_cnt_t` to be of `size_t` within `lmms_basics.h`.  Most of the codebase used `fpp_t` to represent the size of an array of sample frames, which is incorrect, given that `fpp_t` was only 16 bits when sizes greater than 32767 are very possible. `f_cnt_t` was defined as `size_t` as well for consistency.

---------

+++ include/lmms_basics.h
@@ -31,10 +31,7 @@
 #include "lmmsconfig.h"
 
 #include <cstdint>
-#include <array>
 
-#include <cmath>
-#include <algorithm>
 
 
 namespace lmms
@@ -49,8 +46,8 @@ using sample_t = float;		  // standard sample-type
 using int_sample_t = int16_t; // 16-bit-int-sample
 
 using sample_rate_t = uint32_t; // sample-rate
-using fpp_t = int16_t;			// frames per period (0-16384)
-using f_cnt_t = int32_t;		// standard frame-count
+using fpp_t = size_t;			// frames per period (0-16384)
+using f_cnt_t = size_t;			// standard frame-count
 using ch_cnt_t = uint8_t;		// channel-count (0-DEFAULT_CHANNELS)
 using bpm_t = uint16_t;			// tempo (MIN_BPM to MAX_BPM)
 using bitrate_t = uint16_t;		// bitrate in kbps
